test: extract outbound delivery lifecycle coverage
(cherry picked from commit c2a9c5699dca19e5f76346673b263bccafe59873)
